requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2

# ↱ USE THIS IN DEVELOPMENT
psycopg2==2.9.10

psycopg2-binary==2.9.10
# ↳ USE ONLY THIS ONE IN PRODUCTION

python-dotenv==1.0.1
Werkzeug==3.1.3
